requirements.txt
bng_latlon==1.1
Django==4.1.3
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
httplib2==0.21.0
netCDF4==1.6.2
numpy==1.24.0
pandas==1.5.1
pvlive_api==0.12
pyproj==3.4.1
python_dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rest_framework_simplejwt==0.0.1
schedule==1.1.0
setuptools==65.5.0
windpowerlib==0.2.1
git+https://github.com/SheffieldSolar/PV_Live-API@0.12
